+++ test/sql/create_drop.sql
@@ -62,8 +62,9 @@ select emaj.emaj_assign_table('dummySchema','mytbl1','myGroup1');
 select emaj.emaj_assign_table('emaj','mytbl1','myGroup1');
 -- bad table
 select emaj.emaj_assign_table('myschema1','dummyTable','myGroup1');
--- partitionned table (abort for lack of PRIMARY KEY with prior PG versions)
+-- partitionned table (successful with PG9.6- versions)
 select emaj.emaj_assign_table('myschema4','mytblp','myGroup1');
+select emaj.emaj_remove_table('myschema4','mytblp');
 -- temp table
 begin;
   CREATE TEMPORARY TABLE myTempTbl (
@@ -182,9 +183,11 @@ select emaj.emaj_assign_tables('myschema2','','','myGroup2');
 select emaj.emaj_assign_tables('myschema2','mytbl1','mytbl1','myGroup2');
 
 -- excluded tables
--- bad types
+-- bad types (partitionned table is successful with PG9.6- versions)
 select emaj.emaj_assign_tables('myschema4','mytblp$','','myGroup2');
 select emaj.emaj_assign_tables('myschema5','unlogged|oids','','myGroup2');
+-- partitionned table is successful with PG9.6- versions, so remove it
+select emaj.emaj_remove_table('myschema4','mytblp');
 -- the myoidstbl table is removed with PG12+: in these version the myoids table has no OIDS as this propertiy doesn't exist anymore
 select emaj.emaj_remove_table('myschema5','myoidstbl');
 -- temp table
@@ -208,7 +211,7 @@ select emaj.emaj_assign_tables('myschema2','mytbl(5|6)$','','myGroup2');
 select emaj.emaj_assign_tables('myschema2','mytbl(5|6)$','','myGroup2');
 
 -- assign partitions
-select emaj.emaj_assign_tables('myschema4','.*',null,'myGroup4');
+select emaj.emaj_assign_tables('myschema4','.*','mytbl(p|r)$','myGroup4');
 
 -- assign unlogged and withoids tables in an audit_only group
 select emaj.emaj_assign_tables('myschema5','.*',null,'myGroup5');