+++ test/sql/rollback.sql
@@ -376,6 +376,7 @@ select * from emaj.emaj_logged_rollback_group('emptyGroup','EGM3');
 -----------------------------
 -- test use of partitionned tables
 -----------------------------
+select emaj.emaj_assign_table('myschema4', 'mytblr','myGroup4');
 select emaj.emaj_start_group('myGroup4','myGroup4_start');
 insert into myschema4.myTblM values ('2001-09-11',0,'abc'),('2011-09-11',10,'def'),('2021-09-11',20,'ghi');
 insert into myschema4.myTblP values (-1,'abc'),(0,'def'),(1,'ghi');
@@ -383,6 +384,7 @@ insert into myschema4.myTblP values (-1,'abc'),(0,'def'),(1,'ghi');
 select emaj.emaj_set_mark_group('myGroup4','mark1');
 truncate myschema4.myTblM;
 update myschema4.myTblP set col2 = 'DEF' where col1 = 0;
+insert into myschema4.mytblr (col2,col3) values (1, 'abc');
 
 select * from emaj.emaj_logged_rollback_group('myGroup4','mark1');