+++ sql/emaj--4.0.0--4.0.1.sql
@@ -1,5 +1,5 @@
 --
--- E-Maj: migration from 4.0.0 to <devel>
+-- E-Maj: migration from 4.0.0 to 4.0.1
 --
 -- This software is distributed under the GNU General Public License.
 --
@@ -63,7 +63,7 @@ $do$;
 
 -- Insert the upgrade begin record in the operation history.
 INSERT INTO emaj.emaj_hist (hist_function, hist_event, hist_object, hist_wording)
-  VALUES ('EMAJ_INSTALL','BEGIN','E-Maj <devel>', 'Upgrade from 4.0.0 started');
+  VALUES ('EMAJ_INSTALL','BEGIN','E-Maj 4.0.1', 'Upgrade from 4.0.0 started');
 
 -- Lock emaj_group table to avoid any concurrent E-Maj activity.
 LOCK TABLE emaj.emaj_group IN EXCLUSIVE MODE;
@@ -1625,12 +1625,12 @@ INSERT INTO pg_catalog.pg_description (objoid, classoid, objsubid, description)
 
 -- Update the version id in the emaj_param table.
 ALTER TABLE emaj.emaj_param DISABLE TRIGGER emaj_param_change_trg;
-UPDATE emaj.emaj_param SET param_value_text = '<devel>' WHERE param_key = 'emaj_version';
+UPDATE emaj.emaj_param SET param_value_text = '4.0.1' WHERE param_key = 'emaj_version';
 ALTER TABLE emaj.emaj_param ENABLE TRIGGER emaj_param_change_trg;
 
 -- Insert the upgrade end record in the operation history.
 INSERT INTO emaj.emaj_hist (hist_function, hist_event, hist_object, hist_wording)
-  VALUES ('EMAJ_INSTALL','END','E-Maj <devel>', 'Upgrade from 4.0.0 completed');
+  VALUES ('EMAJ_INSTALL','END','E-Maj 4.0.1', 'Upgrade from 4.0.0 completed');
 
 -- Post installation checks.
 DO